Create and add a sidebar with a color selector to change painting color
This sidebar has a button (ColourChangeButton) that opens a CallOutBox
containing a ColourSelector. The sidebar takes a reference of the
DrawingCanvas and passes that reference to the ColourChangeButton so that
it can change the color of the Graphics object that draws the path.

Note:
The drawing's precision, responsiveness, and speed decreases dramatically
as the window gets larger. Also, when the painting color changes, it changes
the color of every subpath. This should be fixed to only change the color of
the current draw.