Fix some (not all) resizing issues with SideBar and DrawingCanvas
When the whole window was resized, SideBar would resize according to
the window size. Now, it has a fixed size and can only be resized from
the ResizableEdgeComponent. Also, fixed a contraint issue with the
ResizableEdgeComponent.

Note: The Sidebar and DrawingCanvas resize back to their original sizes
and positions when the window (or MainContentComponent) gets resized.