Fix resizing and LookAndFeel issues; Use Rectangle for bounds.
Fix issue with resizing where when resizing the window, DrawingCanvas
would keep the size led on by resizing the SideBar with ResizableEdgeComponent.
Most of the resizing code is done inside of SideBar, meaning it handles the
resizing of DrawingCanvas and itself. The MainComponent just calls Sidebar's
resized() to update it when the window is resized.

Rectangle objects have also been added to be used as the component's boundaries.
This is mostly to experiment with using Rectangle and learning how it can be
used.

Note:
More things need to be considered and the whole resize design could be improved,
but it works as expected at the moment. A cleaner, more efficient solution should
be considered.