Refactor the resizing code from earlier commits
Refactor the resizing code and logic so that it's a lot cleaner
and easier to understand. Also, it uses a lot less logic and
redundant function calls.

The resizing of the two components (SideBar and DrawingCanvas) are
still mostly done inside of SideBar::resized(), except for one part
in MainComponent that sets the bounds of the SideBar (which in turn
sets the bounds of DrawingCanvas) relative to the window size.